Custom code to solve utility problems
As part of coding assignments, I've solved problems including, how to pick a most popular celebrity given history file, how to pick projects with best returns, how to identify smallest recurring pattern in a string of characters, & how to read a list of numbers and translate the face value to English words. I had a lot of fun. I'd like to scrub this as a useful reference (notebook) for python.